Create Group 2 Problem 1
The Pythagorean Theorem states that the sum of the squares of the sides of a right triangle is equal to the
square of the hypotenuse. For example, if two sides of a right triangle have lengths 3 and 4, then the
hypotenuse must have a length of 5. The integers 3, 4, and 5 together form a Pythagorean triple. There is an
infinite number of such triples. Design and develop an algorithm as a C++ program that takes two positive
integers, m and n, where m > n, and computes the Pythagorean triple according to the following formulas: